Refactor: Move SplashScreen to designsystem module
- Moved SplashScreen composable from `app` module to `core/designsystem` module.
- Updated Kover configuration to reflect path changes for MainApp.
- Deleted obsolete SplashScreenTest.kt as SplashScreen is now a simpler component.
- Adjusted imports in KmtAppTest, Main.kt (wasmJs), and MainApp.kt (jvm) to use the new SplashScreen location.
- Removed test tags from SplashScreen as they are no longer necessary.
