+++ app/src/jvmMain/kotlin/com/mshdabiola/kmtemplate/MainApp.kt
@@ -42,7 +42,7 @@ import com.mshdabiola.kmtemplate.app.generated.resources.Res
 import com.mshdabiola.kmtemplate.app.generated.resources.desktopicon
 import com.mshdabiola.kmtemplate.di.appModule
 import com.mshdabiola.kmtemplate.ui.KmtApp
-import com.mshdabiola.kmtemplate.ui.SplashScreen
+import com.mshdabiola.designsystem.component.SplashScreen
 import com.mshdabiola.model.CustomLogWriter
 import com.mshdabiola.model.Platform
 import kotlinx.coroutines.delay
