+++ core/designsystem/src/commonMain/kotlin/com/mshdabiola/designsystem/component/SplashScreen.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mshdabiola.kmtemplate.ui
+package com.mshdabiola.designsystem.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
@@ -28,13 +28,11 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag // Ensure this is imported
 import androidx.compose.ui.unit.dp
 import com.mshdabiola.designsystem.drawable.KmtIcons
 import com.mshdabiola.designsystem.strings.KmtStrings
 import com.mshdabiola.designsystem.theme.onPrimaryLight
 import com.mshdabiola.designsystem.theme.primaryLight
-import com.mshdabiola.model.testtag.SplashScreenTestTags
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
@@ -43,8 +41,7 @@ fun SplashScreen(
 ) {
     Surface(
         modifier = modifier
-            .fillMaxSize() // Apply fillMaxSize to the Surface as well for the root
-            .testTag(SplashScreenTestTags.SCREEN_ROOT), // Apply root tag here
+            .fillMaxSize(),
         color = primaryLight,
     ) {
         Column(
@@ -56,15 +53,13 @@ fun SplashScreen(
         ) {
             Image(
                 modifier = Modifier
-                    .sizeIn(maxWidth = 200.dp, maxHeight = 200.dp)
-                    .testTag(SplashScreenTestTags.BRAND_IMAGE), // Tag for the image
+                    .sizeIn(maxWidth = 200.dp, maxHeight = 200.dp),
                 imageVector = KmtIcons.AppIcon,
                 contentDescription = "app icon", // Keep contentDescription for accessibility
             )
 
             Spacer(Modifier.height(32.dp))
             Text(
-                modifier = Modifier.testTag(SplashScreenTestTags.BRAND_TEXT), // Tag for the text
                 text = KmtStrings.brand,
                 style = MaterialTheme.typography.headlineSmall,
                 color = onPrimaryLight,
